+++ app/controllers/pg_hero/home_controller.rb
@@ -111,7 +111,7 @@ module PgHero
 
     def live_queries
       @title = "Live Queries"
-      @running_queries = @database.running_queries
+      @running_queries = @database.running_queries(all: true)
     end
 
     def queries